Re-enables all the buttons when refreshing
Firefox doesn't reset the states of the buttons when refreshing.  As
we disabled some buttons when the user finishes a whole round of
memorizing, if the user hits on "refresh" to reload the page, those
grayed-out buttons will retain disabled.

In order to work around this problem, we enable those buttons during
initialization.